remove inline style in audioplayer q
