+++ package.json
@@ -18,7 +18,8 @@
     "react-i18next": "^11.7.0",
     "react-player": "^2.6.0",
     "react-scripts": "3.4.1",
-    "react-test-renderer": "^16.13.1"
+    "react-test-renderer": "^16.13.1",
+    "sass": "^1.26.10"
   },
   "scripts": {
     "start": "npm run watch:css && react-scripts start",
