+++ src/components/AudioPage/style.scss
@@ -6,3 +6,6 @@ $font-roboto: 'Roboto';
   font-size: 2rem;
   font-family: $font-roboto;
 }
+.height-200 {
+  height: 200px;
+}
